# PERF: DataFrame.iloc[int] for EA dtypes