@jbrockmendel 2023-08-12
Is the issue that the relevant `EA.__setitem__` methods are non-performant?