@lukemanley 2023-08-12
> Is the issue that the relevant `EA.__setitem__` methods are non-performant?

Yes. In the case pyarrow, really non-performant to iteratively set each element.