@jbrockmendel 2023-08-31
I'm not wild about this.  Seems to be papering over a hacky `__setitem__` implementation for the ArrowEA, which really should just be immutable.