@lukemanley 2023-09-01
> I'm not wild about this. Seems to be papering over a hacky `__setitem__` implementation for the ArrowEA, which really should just be immutable.

Fair enough. The `__setitem__` performance for ArrrowEA was pretty rough here. It was copying the new array for every element set. We could revert, but it would reintroduce that behavior. Is there some change you would like to see in the near term? (revert, TODO note, special-case ArrowEA, something else?)